Move SNAX common parts to common base class
