+++ compiler/transforms/convert_linalg_to_acc.py
@@ -12,7 +12,7 @@ from xdsl.pattern_rewriter import (
 )
 from xdsl.traits import SymbolTable
 
-from compiler.accelerators.snax_hwpe_mult import HWPEAcceleratorInfo
+from compiler.accelerators.snax_hwpe_mult import SNAXHWPEMultAcceleratorInterface
 from compiler.dialects import acc
 
 
@@ -34,7 +34,7 @@ class ConvertLinalgToAcceleratorPattern(RewritePattern):
             return
 
         # grab accelerator
-        accelerator = HWPEAcceleratorInfo()
+        accelerator = SNAXHWPEMultAcceleratorInterface()
 
         t = self.module.get_trait(SymbolTable)
         assert t is not None
